feat(julia): add adev wrapper and noise slope tests
- julia/src/deviations.jl: thin adev wrapper over the shared engine
  (kernel + DevParams), following deviation-engine skill architecture
- julia/src/SigmaTau.jl: include deviations.jl, export adev
- julia/test/runtests.jl: adev API smoke test + noise slope validation
  (white PM → τ^-1, white FM → τ^-1/2, ±0.15 tolerance at N=4096)
- julia/Project.toml: move Test/Random to [extras]/[targets] for correct
  test isolation; fix _compute_sz(0.0, …) to avoid Int→Float64 type
  mismatch in calculate_edf when sz0 evaluates to zero

+++ julia/src/stats.jl
@@ -76,7 +76,7 @@ function calculate_edf(alpha::Int, d::Int, m::Int, F::Int, S::Int, N::Int)
     M = 1 + floor(Int, S * (N - L) / m)   # number of summands
     J = min(M, (d + 1) * S)               # truncation parameter (Greenhall 2003)
 
-    sz0        = _compute_sz(0,    F, alpha, d)
+    sz0        = Float64(_compute_sz(0.0, F, alpha, d))   # 0.0 ensures Float64 throughout
     basic_sum  = _compute_basic_sum(J, M, S, F, alpha, d, sz0)
 
     basic_sum > 0 || return NaN
